refactor : 컨트롤러에 있던 로직들을 도메인으로 옮김
