+++ src/main/java/lotto/LottoApplicaion.java
@@ -3,23 +3,18 @@ package lotto;
 import static lotto.view.InputView.*;
 import static lotto.view.ResultView.*;
 
-import java.util.List;
-
-import lotto.controller.Statistics;
-import lotto.controller.TicketGenerator;
-import lotto.controller.WinningNumberChecker;
-import lotto.domain.Lottos;
+import lotto.domain.ProfitCalculator;
 import lotto.domain.Ticket;
 
 public final class LottoApplicaion {
 	public static void main(String[] args) {
 		int price = printPriceInput();
 		int count = printPrice(price);
-		Ticket ticket = new Ticket(TicketGenerator.generate(count));
+		Ticket ticket = new Ticket(Ticket.generate(count));
 		printLottoNumber(ticket);
 		printWinningNumber();
 		int[] winningNumbers = winningNumbers(inputText());
-		int[] statistics = Statistics.calculate(WinningNumberChecker.check(winningNumbers, ticket, count));
-		printStatistics(statistics, Statistics.calculateRateOfReturn(statistics, price));
+		int[] statistics = ProfitCalculator.calculate(ProfitCalculator.check(winningNumbers, ticket, count));
+		printStatistics(statistics, ProfitCalculator.calculateRateOfReturn(statistics, price));
 	}
 }
