--- src/main/java/lotto/controller/TicketGenerator.java
@@ -1,18 +1,0 @@
-package lotto.controller;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import lotto.domain.Lottos;
-
-public final class TicketGenerator {
-	public static List<Lottos> generate(final int number) {
-		List<Lottos> entireLottos = new ArrayList<>();
-
-		for (int i = 0; i < number; i++) {
-			entireLottos.add(new Lottos(LottosGenerator.generate()));
-		}
-
-		return entireLottos;
-	}
-}
