+++ src/test/java/lotto/domain/LottosTest.java
@@ -20,4 +20,10 @@ public final class LottosTest {
 
 		assertThat(lottos.contains(winningNumbers)).isEqualTo(5);
 	}
+
+	@DisplayName("로또 번호 6개 생성.")
+	@Test
+	void 로또_번호_6개_생성() {
+		assertThat(Lottos.generate().get(0).isBetween(1, 45)).isTrue();
+	}
 }
