+++ src/test/java/lotto/domain/ProfitCalculatorTest.java
@@ -1,8 +1,7 @@
-package lotto.controller;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -11,9 +10,19 @@ import org.junit.jupiter.api.Test;
 
 import lotto.domain.Lotto;
 import lotto.domain.Lottos;
+import lotto.domain.ProfitCalculator;
 import lotto.domain.Ticket;
 
-public final class WinningNumberCheckerTest {
+public final class ProfitCalculatorTest {
+	@DisplayName("등수마다 일치하는 로또 개수 반환.")
+	@Test
+	void 등수마다_일치하는_로또_개수() {
+		int[] numbersOfMatch = {1, 2, 3, 4, 5};
+		int[] result = ProfitCalculator.calculate(numbersOfMatch);
+
+		assertThat(result[2] == 1).isTrue();
+	}
+
 	@DisplayName("당첨 번호와 내 번호를 비교해서 일치하는 개수 반환")
 	@Test
 	void 당첨번호와_내번호를_비교해서_일치하는_개수_반환() {
@@ -25,7 +34,7 @@ public final class WinningNumberCheckerTest {
 		List<Lottos> lottosList = List.of(lottos);
 		Ticket ticket = new Ticket(lottosList);
 
-		int[] result = WinningNumberChecker.check(winningNumbers, ticket, number);
+		int[] result = ProfitCalculator.check(winningNumbers, ticket, number);
 
 		assertThat(result[0] == 6).isTrue();
 	}
